refactor: extract models.ts and web-tools.ts from index.ts
